Add onSubmit to wizard step type to facilitate separate submit logic per step
