+++ src/lib/types.ts
@@ -3,4 +3,5 @@ export type stepsType = {
     path: string;
     description: string;
     component: React.FunctionComponentElement<any>;
+    onSubmit?: (result: any) => void;
 }
