+++ src/pages/new-request/items/items-form.tsx
@@ -2,7 +2,6 @@ import { DataTable } from "@/components/data-table/data-table"
 import { columns } from "@/constants/data-table/columns"
 import { sampleItems1 } from "@/constants/fake-data/request-item"
 
-
 const ItemsForm = ({ isStatic = false }: { isStatic?: boolean }) => {
  
   return (
@@ -10,6 +9,10 @@ const ItemsForm = ({ isStatic = false }: { isStatic?: boolean }) => {
       <DataTable columns={columns} data={sampleItems1} isStatic={isStatic}/>
     </div>
   )
+}
+
+export const submitItems = () => {
+  console.log("Submit Items Info")
 }
 
 export default ItemsForm
